requirements.txt
aws-cdk-lib 
PyYAML
requests_aws4auth